refactor: initial code structure update
