+++ src/app/(private)/dashboard/components/financial-dashboard.tsx
@@ -1,6 +1,9 @@
 "use client";
 
-import { checkPinExists, getTransactions } from "@/src/app/dashboard/services";
+import {
+  checkPinExists,
+  getTransactions,
+} from "@/src/app/(private)/dashboard/services";
 import UserPinFormDialogDrawer from "@/src/components/user-pin-form-dialog-drawer";
 import Column from "@/src/components/utils/column";
 import Show from "@/src/components/utils/show";
