+++ src/app/(private)/dashboard/components/financial-summary-card.tsx
