+++ src/app/(private)/minha-conta/page.tsx
