+++ src/app/(public)/entrar/services/do-logout.ts
