--- src/components/top-loading-bar.tsx
@@ -1,25 +1,0 @@
-"use client";
-
-import { usePathname } from "next/navigation";
-import NProgress from "nprogress";
-import { useEffect } from "react";
-
-const TopLoadingBar = () => {
-  const pathname = usePathname();
-
-  useEffect(() => {
-    NProgress.start();
-    const timeout = setTimeout(() => {
-      NProgress.done();
-    }, 300);
-
-    return () => {
-      clearTimeout(timeout);
-      NProgress.done();
-    };
-  }, [pathname]);
-
-  return null;
-};
-
-export default TopLoadingBar;
